fix: score multiplier regression — restore old 1/2/3 values, unify save/restore
The magic-number refactor accidentally changed GameController.score_multiplier
from the original {'easy':1,'medium':2,'hard':3} to the milestone-difficulty
multiplier (1.0, 1.5, 2.0), reducing medium score gain by 25% and hard by 33%.

Fix: update DIFFICULTY_MULTIPLIERS in constants.py to (1.0, 2.0, 3.0) matching
the original score multiplier values (which also correctly scales milestone
thresholds). Also make restore_from_save use the same function instead of
a second hardcoded dict, and fix score_multiplier type hints int→float.

+++ airwar/game/managers/game_controller.py
@@ -25,7 +25,7 @@ class GameState:
     difficulty: str = 'medium'
     username: str = 'Player'
     score: int = 0
-    score_multiplier: int = 1
+    score_multiplier: float = 1.0
     paused: bool = False
     running: bool = True
     player_invincible: bool = False
